+++ custom_components/viewfold/const.py
@@ -3,4 +3,4 @@
 DOMAIN = "viewfold"
 FRONTEND_URL = "/viewfold/viewfold.js"
 CONF_DEBUG = "debug"
-VERSION = "0.1.2"
+VERSION = "0.2.0"