+++ src/config/defaults.ts
@@ -1,11 +1,3 @@
-export const VISUAL_COVER_DEVICE_CLASSES = new Set([
-  "awning",
-  "blind",
-  "curtain",
-  "shade",
-  "shutter",
-]);
-
 export const COVERS_PATH = "covers";
 export const COVERS_ICON = "mdi:window-shutter";
 export const COVERS_COLOR = "blue-grey";